pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "markdown>=3.9",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44",
    "werkzeug>=3.1.3",
]
